+++ R/AFM.lineProfile.R
@@ -5,6 +5,8 @@
 #'   points are provided in units of nanometers or pixels. If the starting
 #'   and end point coordinates are not provided, it will use the \code{raster::click()}
 #'   function to prompt the user to click on two points on the graph.
+#'   
+#'   Note: the convention is that the bottom left corner is (1,1) in pixels and (0,0) in nanometers
 #'
 #' @param obj AFMdata object
 #' @param x1 start x position in  units of nm/pixels from bottom left, if \code{NA}, user will need to click on two points to define profile line
@@ -34,7 +36,8 @@
 AFM.lineProfile <- function(obj,x1=NA,y1=NA,x2=NA,y2=NA,
                             N=1,
                             unitPixels = FALSE, verbose=FALSE) {
-  xD = 0
+  
+  if (unitPixels) xD = 0 else xD = 0 # define left bottom corner as (1,1) pixels and (0,0)nm
   N = min(obj@y.pixels, N)
   for(i in 1:N) {
     obj = .singleLineProfile(obj, x1, y1 + xD, x2, y2 + xD, unitPixels, verbose)
@@ -68,21 +71,8 @@ AFM.lineProfile <- function(obj,x1=NA,y1=NA,x2=NA,y2=NA,
     unitPixels = FALSE
   }
 
-
-  if (!unitPixels) {
-    range.x = max(d$x) - min(d$x)
-    range.y = max(d$y) - min(d$y)
-    if(x1 >= range.x) { warning("x1: Out of range"); x1=0.99*range.x }
-    if(y1 >= range.y) { warning("y1: Out of range"); y1=0.99*range.y }
-    if(x2 >= range.x) { warning("x2: Out of range"); x2=0.99*range.x}
-    if(y2 >= range.y) { warning("y2: Out of range"); y2=0.99*range.y}
-
-    x1.pixel = round(x1/range.x*width.x)
-    y1.pixel = round(y1/range.y*width.y)
-    x2.pixel = round(x2/range.x*width.x)
-    y2.pixel = round(y2/range.y*width.y)
-  } else {
-    if ((x1<0) | (x2<0) | (y1<0) | (y2<0)) { warning("Must have positive coordinates.") }
+  if (unitPixels) { # input is in units of pixels, from (1,1)
+    if ((x1<1) | (x2<1) | (y1<1) | (y2<1)) { warning("Must have non-zero positive pixel coordinates."); x1=min(1,x1); x2=min(1,x2); y1=min(1,y1); y2=min(1,y2); }
     if (x1 > width.x) { warning("x1: Out of range"); x1 = width.x }
     if (x2 > width.x) { warning("x2: Out of range"); x2 = width.x }
     if (y1 > width.y) { warning("y1: Out of range"); y1 = width.y }
@@ -91,6 +81,18 @@ AFM.lineProfile <- function(obj,x1=NA,y1=NA,x2=NA,y2=NA,
     y1.pixel = y1
     x2.pixel = x2
     y2.pixel = y2
+  } else {
+    range.x = AFMcopy@x.conv * AFMcopy@x.pixels # max(d$x) - min(d$x)
+    range.y = AFMcopy@y.conv * AFMcopy@y.pixels # max(d$y) - min(d$y)
+    if(x1 > range.x) { warning("x1: Out of range"); x1=range.x }
+    if(y1 > range.y) { warning("y1: Out of range"); y1=range.y }
+    if(x2 > range.x) { warning("x2: Out of range"); x2=ange.x}
+    if(y2 > range.y) { warning("y2: Out of range"); y2=range.y}
+
+    x1.pixel = ceiling((x1+1)/range.x*width.x)
+    y1.pixel = ceiling((y1+1)/range.y*width.y)
+    x2.pixel = floor((x2+1)/range.x*width.x)
+    y2.pixel = floor((y2+1)/range.y*width.y)
   }
 
   AFMcopy@history <- paste(AFMcopy@history,
@@ -103,7 +105,7 @@ AFM.lineProfile <- function(obj,x1=NA,y1=NA,x2=NA,y2=NA,
   Dy = - abs(y2.pixel - y1.pixel)
   sy = sign(y2.pixel - y1.pixel)
   er = Dx + Dy
-  r = c(x1.pixel + y1.pixel * width.x)
+  r = c(x1.pixel + (y1.pixel-1) * width.x)
   q2=0
   r2 = c(q2)
   # Bresenham's Line Algorithm
@@ -121,7 +123,7 @@ AFM.lineProfile <- function(obj,x1=NA,y1=NA,x2=NA,y2=NA,
       ly = AFMcopy@y.conv
     }
     # add data point
-    q1 = x1.pixel + y1.pixel * width.x
+    q1 = x1.pixel + (y1.pixel-1) * width.x
     q2 = q2+sqrt(lx^2 + ly^2)
     r=c(r, q1)
     r2 = c(r2, q2)